Move UseGraphSonX into Core.
