+++ test/ExRam.Gremlinq.PublicApi.Tests/PublicApiTests.SupportNewtonsoftJson.verified.cs
@@ -3,8 +3,6 @@
     public static class GremlinQueryEnvironmentExtensions
     {
         public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment StoreTimeSpansAsNumbers(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
-        public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment UseGraphSon2(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
-        public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment UseGraphSon3(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
         public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment UseNewtonsoftJson(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
     }
 }
